change the rest clients a bit
