+++ comms-connector/.env.example
@@ -1,10 +1,12 @@
 MYSQL_DATABASE=comms-db
 MYSQL_ROOT_PASSWORD=root
 
-COMM_DISCORD_SECRET=Bot ${discord bot token}
+COMM_DISCORD_SECRET="discord bot token"
 COMM_DISCORD_API_VERSION=v10
 COMM_DISCORD_BASE_URL=https://discord.com/api
 
 TRACE_GENAI_BASE_URL=http://localhost:8080
+
+TRACE_PROJECT_MANAGEMENT_BASE_URL=http://localhost:8080
 
 DEBUG=true
