+++ comms-connector/src/main/resources/application.properties
@@ -14,3 +14,5 @@ trace.discord.api-version=${COMM_DISCORD_API_VERSION}
 trace.discord.base-url=${COMM_DISCORD_BASE_URL}
 
 trace.genai.base-url=${TRACE_GENAI_BASE_URL}
+
+trace.project-management.base-url=${TRACE_PROJECT_MANAGEMENT_BASE_URL}
